+++ WebAppAPI/Entities/ScanMachineContext.cs
@@ -39,7 +39,7 @@ public partial class ScanMachineContext : DbContext
 
     protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
 #warning To protect potentially sensitive information in your connection string, you should move it out of source code. You can avoid scaffolding the connection string by using the Name= syntax to read it from configuration - see https://go.microsoft.com/fwlink/?linkid=2131148. For more guidance on storing connection strings, see https://go.microsoft.com/fwlink/?LinkId=723263.
-        => optionsBuilder.UseSqlServer("Data Source=scanmachine.cxksysgcg0vx.ap-southeast-2.rds.amazonaws.com;Initial Catalog=ScanMachine;User ID=admin;Password=voquanghuy903506;Connect Timeout=30;Encrypt=True;TrustServerCertificate=True;");
+        => optionsBuilder.UseSqlServer("Data Source=LAPTOP-C14F39CJ\\SQLEXPRESS;Initial Catalog=ScanMachine;User ID=sa;Password=12345;Connect Timeout=30;Encrypt=True;TrustServerCertificate=True;");
 
     protected override void OnModelCreating(ModelBuilder modelBuilder)
     {
