+++ WebAppAPI/bin/Debug/net8.0/appsettings.json
@@ -6,7 +6,7 @@
     }
   },
   "ConnectionStrings": {
-    "DBConnetion": "Data Source=scanmachine.cxksysgcg0vx.ap-southeast-2.rds.amazonaws.com;Initial Catalog=ScanMachine;User ID=admin;Password=voquanghuy903506;Connect Timeout=30;Encrypt=True;TrustServerCertificate=True;"
+    "DBConnetion": "Data Source=LAPTOP-C14F39CJ\\SQLEXPRESS;Initial Catalog=ScanMachine;User ID=sa;Password=12345;Connect Timeout=30;Encrypt=True;TrustServerCertificate=True;"
   },
